% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_helpers.R
\name{alphabetQuality}
\alias{alphabetQuality}
\title{This filters out sequences which have bad average quality readings.}
\usage{
alphabetQuality(reads)
}
\arguments{
\item{reads}{(ShortRead object) Loaded reads from fastq.}
}
\value{
(Logical) Logical vector with the valid rows as TRUE.
}
\description{
This filters out sequences which have bad average quality readings.
}

