% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplican.R
\docType{package}
\name{amplican}
\alias{amplican}
\alias{amplican-package}
\title{amplican: A package for something.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{amplican functions}{

The amplican functions ...
}

