% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplican.R
\name{amplican_default}
\alias{amplican_default}
\title{Wraps main package functionality into one function.}
\usage{
amplican_default(config, fastq_folder, results_folder)
}
\arguments{
\item{config}{(string) The path to your configuration file. For example:
\code{system.file("extdata", "config.txt", package = "amplican")}}

\item{fastq_folder}{(string) Path to FASTQ files. If not specified, FASTQ files should be
in the same directory as config file.}

\item{results_folder}{(string) Where do you want your results to be stored. The
program will create files in that folder so make sure you have writing permissions.}
}
\description{
amplican_default is convinient wrapper around all default settings of
amplicanAnalysis, amplicanReport and amplicanSummary. It will generate all
results in the result_folder and also print prepared reports into reports folder.
}

