% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{amplican_insertions_plot}
\alias{amplican_insertions_plot}
\title{This function plots insertions in relation to the amplicon. Top plot is for the forward reads, middle one shows
amplicon sequence, and bottom plot is for reverse reads.}
\usage{
amplican_insertions_plot(alignmentsGRanges, config, id)
}
\arguments{
\item{alignmentsGRanges}{(GRanges object) Loaded alignment information from alignments.csv file.}

\item{config}{(data.frame) Loaded table from config_summary.csv file.}

\item{id}{(string) Name of the ID column from config file.}
}
\value{
(insertions plot) ggplot2 object of mutation plot
}
\description{
This function plots insertions in relation to the amplicon. Top plot is for the forward reads, middle one shows
amplicon sequence, and bottom plot is for reverse reads.
}

