% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{amplican_plot_cuts}
\alias{amplican_plot_cuts}
\title{Plots cuts using ggplot2 and ggbio.}
\usage{
amplican_plot_cuts(alignments, config, id)
}
\arguments{
\item{alignments}{(GRanges object) Loaded alignment information from alignments.csv file.}

\item{config}{(data.frame) Loaded table from config_summary.csv file.}

\item{id}{(string or vector of strings) Name of the ID column from config file or name of multiple IDs if it is
possible to group them. First amplicon will be used as the basis for plot.}
}
\value{
(cuts plot) ggplot2 object of cuts plot
}
\description{
This function plots cuts in relation to the amplicon with distinction for
each ID. Top plot is meta-plot of all cuts combined with frequencies for all reads
in the  amplicon group. Bottom plot shows cuts with facets for each ID.
}
\examples{
#example config
config <- read.csv(system.file("extdata", "config.csv", package = "amplican"))
#example alignments results
alignments <- read.csv(system.file("extdata", "results", "alignments.csv", package = "amplican"))
amplican_plot_cuts(alignments, config, c('ID_1','ID_3'))

}
\seealso{
Other specialized plots: \code{\link{amplican_plot_deletions}},
  \code{\link{amplican_plot_insertions}},
  \code{\link{amplican_plot_mismatches}}
}

