% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{amplican_plot_deletions}
\alias{amplican_plot_deletions}
\title{This function plots deletions in relation to the amplicon. Top plot is for the forward reads, middle one shows
amplicon sequence, and bottom plot is for reverse reads.}
\usage{
amplican_plot_deletions(alignments, config, id, cut_buffer = 5)
}
\arguments{
\item{alignments}{(GRanges object) Loaded alignment information from alignments.csv file.}

\item{config}{(data.frame) Loaded table from config_summary.csv file.}

\item{id}{(string) Name of the ID column from config file.}

\item{cut_buffer}{(numeric) Default is 5, you should specify the same as used in the analysis.}
}
\value{
(deletions plot) ggplot2 object of mutation plot
}
\description{
This function plots deletions in relation to the amplicon. Top plot is for the forward reads, middle one shows
amplicon sequence, and bottom plot is for reverse reads.
}

