% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{amplican_plot_insertions}
\alias{amplican_plot_insertions}
\title{Plots insertions using ggplot2 and ggbio.}
\usage{
amplican_plot_insertions(alignments, config, id, cut_buffer = 5,
  xlab_spacing = 4, filter = TRUE)
}
\arguments{
\item{alignments}{(GRanges object) Loaded alignment information from
alignments_events.csv file.}

\item{config}{(data.frame) Loaded table from config_summary.csv file.}

\item{id}{(string or vector of strings) Name of the ID column from config
file or name of multiple IDs if it is
possible to group them. First amplicon will be used as the basis for plot.}

\item{cut_buffer}{(numeric) Default is 5, you should specify the same as
used in the analysis.}

\item{xlab_spacing}{(numeric) Spacing of the x axis labels. Default is 4.}

\item{filter}{(boolean) Whether deletions overlapping primers should be
removed. By deafult it set to TRUE.}
}
\value{
(insertions plot) ggplot2 object of insertions plot
}
\description{
This function plots insertions in relation to the amplicon. Top plot is for
the forward reads, middle one shows
amplicon sequence, and bottom plot is for reverse reads.
}
\examples{
#example config
config <- read.csv(system.file("extdata", "config.csv", package = "amplican"))
#example alignments results
alignments_file <- system.file("extdata", "results", "alignments_events.csv", package = "amplican")
alignments <- read.csv(alignments_file)
amplican_plot_insertions(alignments, config, c('ID_1','ID_3'), 5)

}
\seealso{
Other specialized plots: \code{\link{amplican_plot_cuts}},
  \code{\link{amplican_plot_deletions}},
  \code{\link{amplican_plot_mismatches}}
}
