% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanSummarize.R
\name{amplicanSummarizeHDR}
\alias{amplicanSummarizeHDR}
\title{Summarize how many reads comes from HDR.}
\usage{
amplicanSummarizeHDR(aln, cfgT, HDR = "HDR")
}
\arguments{
\item{aln}{(data.frame) Contains events from the alignments.}

\item{cfgT}{(data.frame) Config file with the experiments details.}

\item{HDR}{(character) Name of the column from aln that contains HDR
information.}
}
\value{
(data.frame) As cfgT, but with extra columns.
}
\description{
Before using this function make sure events are filtered to represent
consensus with \code{amplicanConsensus}, if you use both forward and
reverse reads. If you want to calculate metrics over expected cut site,
filter events using \code{amplicanOverlap}.
}
\details{
Adds columns to cfgT:
\itemize{
\item{HDR}{ Count of reads with HDR events}
}
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "events_filtered_shifted_normalized.csv",
                         package = "amplican")
aln <- data.table::fread(file_path)
aeSet <- readRDS(system.file("extdata", "results", "alignments",
                             "AlignmentsExperimentSet.rds",
                             package = "amplican"))
cfgT <- data.table::fread(
  system.file("extdata", "results", "config_summary.csv",
              package = "amplican"))
# check which events come from HDR, in the example amplicons are donor
# templates so we expect no events
aln$HDR <- amplicanHDR(aln, aeSet, cfgT, cfgT$Amplicon)
# finally calculate HDR counts per experiment
amplicanSummarizeHDR(aln, cfgT)

}
\seealso{
Other analysis steps: \code{\link{amplicanAlign}},
  \code{\link{amplicanConsensus}},
  \code{\link{amplicanFilter}}, \code{\link{amplicanHDR}},
  \code{\link{amplicanMap}},
  \code{\link{amplicanNormalize}},
  \code{\link{amplicanOverlap}},
  \code{\link{amplicanPipeline}},
  \code{\link{amplicanReport}},
  \code{\link{amplicanSummarize}}
}
