% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{ampliconIntegrityCheck}
\alias{ampliconIntegrityCheck}
\title{This function checks if amplicons can be ploted together, aka they are the
same.}
\usage{
ampliconIntegrityCheck(configTable, id)
}
\arguments{
\item{configTable}{(data.frame) data frame of config file}

\item{id}{(vector of characters) vector of IDs that are to be tested}
}
\value{
(boolean) TRUE, If anything goes wrong stops and prints error.
}
\description{
This function checks if amplicons can be ploted together, aka they are the
same.
}

