% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{assignedCount}
\alias{assignedCount}
\title{Get count of assigned reads.}
\usage{
assignedCount(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(numeric)
}
\description{
Get count of assigned reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
writeAlignments(aln, file.path(tempdir(), "aln.txt"))

}
\keyword{internal}
