% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{barcodeData<-}
\alias{barcodeData<-}
\title{Barcode data.}
\usage{
barcodeData(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(data.frame)}
}
\value{
(AlignmentsExperimentSet)
}
\description{
Set barcode data.frame with information on the barcode level.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
barcodeData(aln) <- barcodeData(aln) #replace with the same values as before
}
\keyword{internal}
