% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{checkFileAccess}
\alias{checkFileAccess}
\title{This function check if the given file exist and can be read.}
\usage{
checkFileAccess(filePath)
}
\arguments{
\item{filePath:}{(String) A string the path to the file.}
}
\value{
(Bool) TRUE if we have file access. This means:
         -- If the file path is not NULL.
         -- If the file path is not NA.
         -- If the file path is not "Null", "null", "na", ...
         -- If the file path have read permissions.

         FALSE otherwise.
}
\description{
This function check if the given file exist and can be read.
}

