% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{checkFileWriteAccess}
\alias{checkFileWriteAccess}
\title{This function checks if the given directory exist and can be written to.}
\usage{
checkFileWriteAccess(filePath)
}
\arguments{
\item{filePath:}{(String) A string the path to the file.}
}
\value{
(Bool) TRUE if we have file write access.
Stop if otherwise.
}
\description{
This function checks if the given directory exist and can be written to.
}

