% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{checkReady}
\alias{checkReady}
\title{Gives a config file path and check if exist, is readable, etc...}
\usage{
checkReady(configFilePath)
}
\arguments{
\item{configFilePath}{(string) A string with the absolute path pointing to the config file}
}
\value{
bool, TRUE:  Everything went good.
         FALSE: If the config file doesn't exist or if we don't have read access
                If the string provided was 'NULL', 'NA', '0', 'null' or 'na'
}
\description{
Gives a config file path and check if exist, is readable, etc...
}

