% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{checkTarget}
\alias{checkTarget}
\title{This function checks if the target RNA is in the amplicon.}
\usage{
checkTarget(targetPrimer, amplicon, ID, barcode)
}
\arguments{
\item{targetPrimer}{(string) A sequence of nucleotides in a string format
representing the target}

\item{amplicon}{(string) A sequence of nucleotides in a string format
representing the amplicon}

\item{ID}{(string) The ID from where this target and amplicon came.}

\item{barcode}{(string) The barcode from where this target and amplicon came.}
}
\value{
(string) NULL when target is in amplicon, warning string otherwise.
}
\description{
This function checks if the target RNA is in the amplicon.
}

