% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_helpers.R
\name{countUppercaseGroups}
\alias{countUppercaseGroups}
\title{For a given string, detect how many groups of uppercases are there, where are
they, and how long are they.}
\usage{
countUppercaseGroups(candidate)
}
\arguments{
\item{candidate:}{(String) A string with the nucleotide sequence.}
}
\value{
(List) A list with three element. The first element is an integer with the
         total of upper case group; lets say is X. The second element is a
         list of integer of size X with the start position of each group. The
         third element is a list of integer with the wide of each group.
}
\description{
For example:
}
\details{
asdkfaAGASDGAsjaeuradAFDSfasfjaeiorAuaoeurasjfasdhfashTTSfajeiasjsf

Has 4 groups of uppercases, and they have length 7,4,1, and 3 respectevily.
The function returns a list with the length of each one. So the length of the
list represent.
The function returns a list with 3 elements:
-- How many groups we found. If will be 1 if we found nothing.
-- The starting position for each AP. It would be -1 if we found nothing.
-- The wide for each AP. It will be 1 if we found nothing.
}

