% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{cumsumw}
\alias{cumsumw}
\title{Cumulative sum to calculate shift}
\usage{
cumsumw(x)
}
\arguments{
\item{x}{(\code{\link[IRanges]{IRanges}})}
}
\value{
(numeric vector)
}
\description{
Cumulative sum to calculate shift
}
\keyword{internal}
