% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{decode}
\alias{decode}
\title{Get codons for given string - translate}
\usage{
decode(x)
}
\arguments{
\item{x}{(string)}
}
\value{
(string) codons
}
\description{
Get codons for given string - translate
}
\keyword{internal}
