% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{defGR}
\alias{defGR}
\title{Helper to construct GRanges with additional metadata columns.}
\usage{
defGR(
  x,
  ID,
  score,
  strand_info = "+",
  type = "deletion",
  originally = "",
  replacement = ""
)
}
\arguments{
\item{x}{(\code{\link[IRanges]{IRanges}}) names(x) indicating read_id}

\item{ID}{(string)}

\item{score}{(numeric) scores from the alignments}

\item{strand_info}{(string) Either '+', '-'}

\item{type}{(string)}

\item{originally}{(string) Base pairs on the amplicon.}

\item{replacement}{(string) Base pairs on the read.}
}
\value{
(\code{\link[GenomicRanges]{GRanges}}) Object with meta-data
}
\description{
Helper to construct GRanges with additional metadata columns.
}
\keyword{internal}
