% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{deleteFiles}
\alias{deleteFiles}
\title{This funcion takes a txt file, with a file path per row, and delete all those files.}
\usage{
deleteFiles(listOfFilesFileName, deleteSource = FALSE)
}
\arguments{
\item{listOfFilesFileName:}{(string) Path to the file that containts 0 or more files paths that will be deleted}

\item{deleteSource:}{(bool)   If you want that the file listOfFilesFileName to be ALSO deleted, set this
to TRUE. Default is FALSE}
}
\value{
(int) How many files where deleted. This could be less than the orinal files, as it will only delete
         files that you have write privilege.
}
\description{
This funcion takes a txt file, with a file path per row, and delete all those files.
}

