% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{divideWork}
\alias{divideWork}
\title{This function takes a dataframe and a number of processors and gives back the
same dataframe divided into several subsets. Every subset crafted so each
processors takes aproximatly the same amount of rows. Note that that doesn't
mean that they will take the same workload. This is just a arbitrary way of
dividing the work.}
\usage{
divideWork(totalProcessors, configDataframe)
}
\arguments{
\item{totalProcessors:}{(Int)       An integer with the number of processors.
This is also the amount of groups that will
result from this function. The number must be
1 or bigger.}

\item{configDataframe:}{(Dataframe) A dataframe you want to divide. It doesn't
require any special column. But it must not be
empty. It can be full of NA rows thought.}
}
\value{
(List) A list with several dataframes. Each dataframe is a subgroup of
         configDataframe. If the number of rows of configDataframe is bigger
         than totalProcessors, the list have length totalProcessors. If it is
         smaller, the length is equal to the number of rows of
         configDataframe.
}
\description{
See:
}
\details{
divideWorkBySize()
  divideWorkByBarcode()
}

