% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{divideWorkByBarcode}
\alias{divideWorkByBarcode}
\title{This function takes in a config dataframe and divide the work into barcodes.
This means that instead of dividing the work by the number of sequences per
line as in divideWorkBySize, it will consider the barcode group as a whole,
and give packages based on barcodes.}
\usage{
divideWorkByBarcode(totalProcessors, configDataframe, configFilePath)
}
\arguments{
\item{totalProcessors:}{(Int)       An integer with the number of processors.
This is also the amount of groups that will
result from this function. The number must be
1 or bigger.}

\item{configDataframe:}{(Dataframe) A dataframe you want to divide. It must not be
empty.}

\item{configFilePath:}{(String)    The config file from where the configDataframe
was generated. The dataframe contain the
forward and reverse files paths. If it is not
the absolute path, we need to find where are
the files using this path.}
}
\value{
(List) A list with several dataframes. Each dataframe is a subgroup of
         configDataframe. If the number of rows of configDataframe is bigger
         than totalProcessors, the list have length totalProcessors. If it is
         smaller, the length is equal to the number of rows of
         configDataframe.
}
\description{
That way, we don't need to ddply afterwards many times, only once. Each
barcode have a unique combination of files from where it gets its reads.
}
\details{
See:

divideWork()
  divideWorkBySize()
}

