% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_helpers.R
\name{divideWorkBySize}
\alias{divideWorkBySize}
\title{This function divide the dataframe into similar dataframes based on the
colName values. The function try to divide the work so each processor gets
aproximaly the same amount of colName summations. This is commonly known as
the PARTITION PROBLEM. This implementation uses a greedy approach with order
O(nlogn).}
\usage{
divideWorkBySize(totalProcessors, configDataframe, colName)
}
\arguments{
\item{totalProcessors:}{(Int)       An integer with the number of processors.
This is also the amount of groups that will
result from this function. The number must be
1 or bigger.}

\item{configDataframe:}{(Dataframe) A dataframe you want to divide. It must not be
empty.}

\item{colName:}{(String)    The name of a column in configDataframe. This
column should be full of of values that can be
sorted and can be added (typically either int
or float). It can be negative values.}
}
\value{
(List) A list with several dataframes. Each dataframe is a subgroup of
         configDataframe. If the number of rows of configDataframe is bigger
         than totalProcessors, the list have length totalProcessors. If it is
         smaller, the length is equal to the number of rows of
         configDataframe.
}
\description{
See:
}
\details{
divideWork()
  divideWorkByBarcode()
}

