% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{experimentData<-}
\alias{experimentData<-}
\title{Experiment data.}
\usage{
experimentData(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(data.frame)}
}
\value{
(AlignmentsExperimentSet)
}
\description{
Set experiment data.frame with information on the experiment level.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
experimentData(aln) <- experimentData(aln) # replace with the same values
}
\keyword{internal}
