% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters_helpers.R
\name{filterByNucleotides}
\alias{filterByNucleotides}
\title{This function filter out sequences that has undefined nucleotides.}
\usage{
filterByNucleotides(forward.df, reverse.df, validRows)
}
\arguments{
\item{Forward.df:}{(Dataframe) Forward dataframe. This dataframes must have a
column named 'Sequence' where the sequence should be.}

\item{Reverse.df:}{(Dataframe) Reverse dataframe. This dataframes must have a
column named 'Sequence' where the sequence should be.}

\item{validRows:}{(Bool[])    Array with bools telling which sequences do you
want to keep (usually all TRUE). Each one corresponde to a row in the dataframes in the
same order. If the value is TRUE it means that we keep it, if the filter pass. If the value is
false, it means that we discard it, regardless of the result of the filter.}
}
\value{
(Bool[]) Array with bools, with the non valid rows set to FALSE.
}
\description{
For each forward and reverse read the function looks for unvalid nucleotides.
If either the forward or reverse read has a unvalid nucleotide, both of them
are considered unvalid reads, hence rendering that row in the dataframe
unused.
}

