% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{filterEOP}
\alias{filterEOP}
\title{Filter Events Overlapping Primers. Message user when many of the events are
filtered.}
\usage{
filterEOP(idRanges, frPrimer, rwPrimer, amplicon)
}
\arguments{
\item{idRanges}{(data.frame) Contains events.}

\item{frPrimer}{(character) forward primer}

\item{rwPrimer}{(character) reverse primer}

\item{amplicon}{(character) amplicon sequence}
}
\value{
(data.frame) filtered data frame of events
}
\description{
Filter Events Overlapping Primers. Message user when many of the events are
filtered.
}

