% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{flipRanges}
\alias{flipRanges}
\title{Reverse complement events that have amplicons with direction 1.}
\usage{
flipRanges(idRanges, configTable)
}
\arguments{
\item{idRanges}{(data.frame) Loaded events.}

\item{configTable}{(data.frame) Loaded configuration file.}
}
\value{
(data.frame) Returns input idRanges, but events for amplicons with
direction 1 reverse complemented, "+" and "-" swapped.
}
\description{
Reverse complement events that have amplicons with direction 1.
}
