% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanSummarize.R
\name{frameshifted_reads_by_ID}
\alias{frameshifted_reads_by_ID}
\title{Summarize counts of reads with frameshifts for each experiment}
\usage{
frameshifted_reads_by_ID(aln, paired_end = TRUE)
}
\arguments{
\item{aln}{(data.frame)}

\item{paired_end}{(boolean)}
}
\value{
(data.frame)
}
\description{
Summarize counts of reads with frameshifts for each experiment
}
