% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{get_amplicon}
\alias{get_amplicon}
\title{amplicon sequence, reverse complemented when needed}
\usage{
get_amplicon(config, id)
}
\arguments{
\item{config}{(data.frame) config table}

\item{id}{(vector) a vector of id's}
}
\value{
(character) amplicon sequence, reverse complemented if Direction 1
}
\description{
amplicon sequence, reverse complemented when needed
}
\keyword{internal}
