% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{getConfigFolder}
\alias{getConfigFolder}
\title{Given a config file path, get the folder containing that config file}
\usage{
getConfigFolder(configFilePath)
}
\arguments{
\item{string}{configFilePath: A string with the absolute path pointing to the config file}
}
\value{
string :  A string with the name of the folder containing the config file
}
\description{
/home/myUser/myGenesData/experiment.txt
}
\details{
The function returns:
myGenesData
}

