% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{getConfigName}
\alias{getConfigName}
\title{Given a config file path, get the name of the config file. For example, for the path:}
\usage{
getConfigName(configFilePath)
}
\arguments{
\item{string}{configFilePath: A string with the absolute path pointing to the config file}
}
\value{
string :  A string with the name of the config file
}
\description{
/home/myUser/myGenesData/experiment.txt
}
\details{
The function returns:
experiment.txt
}

