% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{getEventInfo}
\alias{getEventInfo}
\title{This function takes alignments and gives back the events coordinates.}
\usage{
getEventInfo(align, ID, ampl_shift, strand_info = "+")
}
\arguments{
\item{align}{(\code{\link[pwalign]{PairwiseAlignmentsSingleSubject}})}

\item{ID}{(string)}

\item{ampl_shift}{(numeric vector) Shift events additionally by this value.
PairwiseAlignmentsSingleSubject returns truncated alignments.}

\item{strand_info}{(string) Either '+', '-' or default '*'}
}
\value{
(\code{\link[GenomicRanges]{GRanges}}) Object with meta-data for insertion,
deletion, mismatch
}
\description{
This function takes alignments and gives back the events coordinates.
}
\keyword{internal}
