% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{getReadsFile}
\alias{getReadsFile}
\title{This function get a reads file and put into a dataframe.}
\usage{
getReadsFile(fileName, TEMPFOLDER = NULL, tempFileConn = NULL,
  cropLeft = 0, cropRight = 0)
}
\arguments{
\item{fileName:}{(String) A String with the absolute path to the file.}

\item{TEMPFOLDER:}{(String) Optional, a String with the absolute path to a folder.
In this folder is where the files are going to be unziped.
If nothing is specify, the files will be unziped in the same
folder where the zipped files are.}

\item{tempFileConn:}{(File) Optional, a file where we write the path of the temporal file.
If a file is unzipped, we might want to consider it a temoporal file
and delete it later. In here we keep track of it.}
}
\value{
If there is no errors:
  (Dataframe) A dataframe with the following columns:

      [01] Sequence line
      [02] Quality

  If the FASTQ file has a number of files that is not divisible by 4:
  NULL
}
\description{
The files are usually stored in a .gz file. If the .gz file has been unzip
before the function won't try to unzip it, it will read the .fastq instead.
If the file is a .gz, and there is no .fastq in the same folder with the same
name, the function will unzip the .gz file and leave the .fastq in the same
folder.
}
\details{
Prerrequisites:

The file must exist, and you must have reads rights.

If the file is zipped, you must have write rights for the writing folder.
}

