% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanSummarize.R
\name{hdr_read_ids}
\alias{hdr_read_ids}
\title{Realign reads to the donor template, report those read ids that have better
score against the donor template and are perfect alignments in regions of the
donor template differences toward amplicon sequence.}
\usage{
hdr_read_ids(reads, donor, hdr_events, scoring_matrix, gapOpening = 25,
  gapExtension = 0, type = "overlap")
}
\arguments{
\item{reads}{(PairwiseAlignmentsSingleSubject)}

\item{donor}{(character)}

\item{hdr_events}{(IRanges) events from alignment amplican to donor without
shifting and flipping}

\item{scoring_matrix}{Use the same as in amplicanPipeline.}

\item{gapOpening}{(numeric)}

\item{gapExtension}{(numeric)}

\item{type}{(character) Same as in pairwiseAlignment.}
}
\value{
(numeric) Read ids from reads that are HDR.
}
\description{
Realign reads to the donor template, report those read ids that have better
score against the donor template and are perfect alignments in regions of the
donor template differences toward amplicon sequence.
}
