% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_alignment.R
\name{is_hdr_strict_by_id}
\alias{is_hdr_strict_by_id}
\title{Figure out which reads conform to the HDR using the donor.}
\usage{
is_hdr_strict_by_id(
  aln_id,
  amplicon,
  donor,
  scoring_matrix,
  gap_opening = 25,
  gap_extension = 0,
  donor_mismatch = 3
)
}
\arguments{
\item{aln_id}{(data.table) This are events that contain already consensus column,
they are not shifted yet. These are only for one specific ID.}

\item{amplicon}{(string) amplicon sequence of this ID}

\item{donor}{(string) donor sequence of this ID}

\item{scoring_matrix}{(scoring matrix)}

\item{gap_opening}{(integer)}

\item{gap_extension}{(integer)}

\item{donor_mismatch}{(integer) How many mismatches are allowed in addition to the HDR events?}
}
\value{
(boolean vector) of TRUE where HDR was detected
}
\description{
This is strict detection as compared to `is_hdr` which was designed to be
less specific and allow for all kinds of donors. This method requires that
you have exactly the same events (mismatches, insertions, deletions) as the difference
between amplicon and donor sequences. Additional mismatches are also allowed, but restricted to
only `donor_mismatch` number. This method assumes you have only consensus events and specific ID
only.
}
