% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{make_experiment_rmd}
\alias{make_experiment_rmd}
\title{Make contents for .Rmd file aggregated on Type of Experiment}
\usage{
make_experiment_rmd(results_folder)
}
\arguments{
\item{results_folder}{path to results as string}
}
\value{
c() collection of strings to put into rmd file
}
\description{
Make contents for .Rmd file aggregated on Type of Experiment
}

