% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{make_id_rmd}
\alias{make_id_rmd}
\title{Make contents for .Rmd file aggregated on ID.}
\usage{
make_id_rmd(results_folder, cut_buffer = 5)
}
\arguments{
\item{results_folder}{(string) path to results}

\item{cut_buffer}{(numeric)}
}
\value{
c() collection of strings to put into rmd file
}
\description{
Make contents for .Rmd file aggregated on ID.
}

