% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{make_summary_rmd}
\alias{make_summary_rmd}
\title{Prepare summary report as .Rmd file.}
\usage{
make_summary_rmd(results_folder)
}
\arguments{
\item{results_folder}{(string) Folder containing results from the \code{\link{amplicanAlign}} function,
do not change names of the files.}
}
\value{
(string) contents to write to filess
}
\description{
Prepare summary report as .Rmd file.
}

