% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{map_to_relative}
\alias{map_to_relative}
\title{Map events to their respective relative coordinates specified with
UPPER case.}
\usage{
map_to_relative(aln, cfgT)
}
\arguments{
\item{aln}{(data.frame) List of events to map to the relative coordinates.}

\item{cfgT}{(data.frame) config table}
}
\value{
(GRanges) Same as events, but the coordinates are relative to the
}
\description{
Translate coordinates of GRanges events so that they can be relative to the
amplicon. As point zero we assume first left sided UPPER case letter in the
amplicon.
}
\examples{
# example config
config <- read.csv(system.file("extdata", "config.csv",
                   package = "amplican"))
# example events
events <- read.csv(system.file("extdata", "results", "alignments",
                   "raw_events.csv", package = "amplican"))
# make events relative to the UPPER case
map_to_relative(events, config)

}
\seealso{
Other analysis steps: \code{\link{amplicanAlign}},
  \code{\link{amplicanConsensus}},
  \code{\link{amplicanFilter}},
  \code{\link{amplicanNormalize}},
  \code{\link{amplicanOverlap}},
  \code{\link{amplicanPipeline}},
  \code{\link{amplicanReport}},
  \code{\link{amplicanSummarize}}
}
