% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{plot_amplicon}
\alias{plot_amplicon}
\title{Plots amplicon sequence using ggplot2.}
\usage{
plot_amplicon(amplicon)
}
\arguments{
\item{amplicon}{(character) Sequence of the amplicon to plot.}
}
\value{
(amplicon plot) ggplot2 object of amplicon plot
}
\description{
Plots amplicon sequence using ggplot2.
}
