% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{plot_height}
\alias{plot_height}
\title{Get figure height in inches for number of elements on y axis.}
\usage{
plot_height(x)
}
\arguments{
\item{x}{(numeric) number of elements to fit onto height axis}
}
\value{
(numeric) In inches
}
\description{
Helper function to calculate figure height
based on number of elements to plot for automating sizes of figures in
knited reports.
}
\examples{
plot_height(20)

}
