% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{readCounts<-}
\alias{readCounts<-}
\title{Alignments for forward reads.}
\usage{
readCounts(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(list) Named (experiment IDs) list with elements of}
}
\value{
(AlignmentsExperimentSet)
\code{\link[pwalign]{PairwiseAlignmentsSingleSubject}} class.
}
\description{
Set alignments for forward reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
readCounts(aln) <- readCounts(aln) # replace with the same values
}
\keyword{internal}
