% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{readCounts}
\alias{readCounts}
\title{Alignments for forward reads.}
\usage{
readCounts(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(listOrNULL)
}
\description{
Set alignments for forward reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
readCounts(aln)
}
\keyword{internal}
