% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{revC}
\alias{revC}
\title{Reverse complement string or vector of strings.}
\usage{
revC(toRevComp)
}
\arguments{
\item{toRevComp}{(vector of characters) Can be singular character.}
}
\value{
(vector of characters)
}
\description{
Reverse complement string or vector of strings.
}

