% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{revComp}
\alias{revComp}
\title{Reverse and complement given string or list of strings}
\usage{
revComp(x)
}
\arguments{
\item{x}{(string or vector of strings)}
}
\value{
(string or vector of strings) reverse complemented input
}
\description{
Reverse and complement given string or list of strings
}
\keyword{internal}
