% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_helpers.R
\name{reverseComplement}
\alias{reverseComplement}
\title{This function return the reverse complementary dna string of a given dna
string.
The function respect uppercases. Any other character that is not
a,c,g,t,A,C,G,T is ignored.}
\usage{
reverseComplement(dna)
}
\arguments{
\item{dna}{(String) A string with the nucleotide sequence.}
}
\value{
(String) A string with the nucleotides reversed and complemented
}
\description{
This function return the reverse complementary dna string of a given dna
string.
The function respect uppercases. Any other character that is not
a,c,g,t,A,C,G,T is ignored.
}

