% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{rveReadsType<-}
\alias{rveReadsType<-}
\title{Read type for reverse reads.}
\usage{
rveReadsType(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(list) Named (experiment IDs) list with elements of}
}
\value{
(AlignmentsExperimentSet)
\code{\link[pwalign]{PairwiseAlignmentsSingleSubject}} class.
}
\description{
Set read type for reverse reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
rveReadsType(aln) <- rveReadsType(aln) # replace with the same values
}
\keyword{internal}
