% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{shift_to_subj}
\alias{shift_to_subj}
\title{Make insertions and deletions relative to the full subject sequence}
\usage{
shift_to_subj(x, ampl_shift, subject, strand_info)
}
\arguments{
\item{x}{(IRangesList)}

\item{ampl_shift}{(numeric vector)}

\item{subject}{(vector of strings) as.character(subject(align))}

\item{strand_info}{(string) + or -}
}
\value{
(IRangesList) shifted
}
\description{
Make insertions and deletions relative to the full subject sequence
}
