% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_filters.R
\name{thresholdNEvents}
\alias{thresholdNEvents}
\title{Find Off-targets and Fragmented alignments from reads.}
\usage{
thresholdNEvents(x)
}
\arguments{
\item{x}{(numeric) Normalized by length (and strand) number of events in each
read. Should work on reads from single experiment.}
}
\value{
(numeric) Threshold value for suggested cutoff on the number of
normalized events of the alignments. Values above this threshold are
potential off-targets or low quality alignments.
}
\description{
Will try to detect off-targets and low quality alignments (outliers). It
checks for continuity of normalized number of events and has no
assumptions about underlying distributions. When there is gap in events that
is bigger than expected, will return threshold. When there is less than 1000
scores in \code{x} maximum event number is returned - filter nothing.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "raw_events.csv", package = "amplican")
aln <- data.table::fread(file_path)
aln <- aln_id <- aln[seqnames == "ID_1"] # for first experiment
# get number of events normalized by strand and read length
aln_id <- aln_id[, list(events = (.N/length(unique(strand)))/max(end),
                        counts = max(counts)), by = "read_id"]
threshold <- thresholdNEvents(aln_id$events)
# use threshold to filter...

}
\seealso{
\code{\link{findPD}} \code{\link{findEOP}}

Other filters: \code{\link{findEOP}}, \code{\link{findPD}}
}
