% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_filters.R
\name{thresholdScores}
\alias{thresholdScores}
\title{Find Off-targets and Fragmented alignments from reads.}
\usage{
thresholdScores(x)
}
\arguments{
\item{x}{(numeric) Input alignment scores of all reads. Should work on
scores from single experiment.}
}
\value{
(numeric) Threshold value for suggested cutoff on the scores of the
alignments. Values below this threshold are potential off-targets or low
quality alignments.
}
\description{
Will try to detect off-targets and low quality alignments (outliers). It
performs hierarchical clustering for parameterized Gaussian mixture models
using \code{\link[mclust]{Mclust}}, then calcualtes medians of predicted
clusters. Based on medians selects cluster with the lowest score. Then
threshold is returned as the most sensitive cutoff for that cluster, minimum
score from closest higher score cluster. When there is less than 1000 scores
in \code{x} or clustering will return only one cluster minimum score is
returned.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "raw_events.csv", package = "amplican")
aln <- data.table::fread(file_path)
thresholdScores(aln$score[aln$seqnames == "ID_1"])

}
\seealso{
\code{\link{findPD}} \code{\link{findEOP}}

Other filters: \code{\link{findEOP}}, \code{\link{findPD}}
}
