% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{unassignedCount}
\alias{unassignedCount}
\title{Get count of unassigned reads.}
\usage{
unassignedCount(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(numeric)
}
\description{
Get count of unassigned reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "temp", "barcode_1_aln.rds", package = "amplican")
aln <- readRDS(file_path)
unassignedCount(aln)
}
\keyword{internal}
