% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_directory.R
\name{unpackFastq}
\alias{unpackFastq}
\title{Unpack fastq files if needed, correct paths to the files in configTable.}
\usage{
unpackFastq(configTable, temp_folder)
}
\arguments{
\item{configTable}{(data.frame) Contains configuration file.}

\item{temp_folder}{(string) Where to store unzipped files.}
}
\value{
(data.frame) configTable with Forward_File and Reverse_File updated
if unpacking files was required.
}
\description{
Unpack fastq files if needed, correct paths to the files in configTable.
}

