% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{upperGroups}
\alias{upperGroups}
\title{Detect uppercases as ranges object.}
\usage{
upperGroups(candidate)
}
\arguments{
\item{candidate}{(string) A string with the nucleotide sequence.}
}
\value{
(\code{\link[IRanges]{IRanges}}) A IRanges object with uppercases groups for
given candidate string
}
\description{
For a given string, detect how many groups of uppercases is inside, where
are they, and how long they are.
}
\details{
For example:
   asdkfaAGASDGAsjaeuradAFDSfasfjaeiorAuaoeurasjfasdhfashTTSfajeiasjsf

Has 4 groups of uppercases of length 7, 4, 1 and 3.
}
\keyword{internal}
