% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_alignment_metaplots}
\alias{write_alignment_metaplots}
\title{Writes alignment MetaPlots}
\usage{
write_alignment_metaplots(group, selection, title)
}
\arguments{
\item{group}{(string)}

\item{selection}{(string or vector of strings)}

\item{title}{(string) title for the alignment plots}
}
\value{
string to include in rmd file
}
\description{
Writes alignment MetaPlots
}
