% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_alignment_plots}
\alias{write_alignment_plots}
\title{Writes alignment plots: deltion, insertion and mismatch}
\usage{
write_alignment_plots(ID, cut_buffer = 5)
}
\arguments{
\item{ID}{is of alignment}

\item{cut_buffer}{(numeric) cut detection box will be extended left and right by box_buffer}
}
\value{
string to include in rmd file
}
\description{
Writes alignment plots: deltion, insertion and mismatch
}

