% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_explanation}
\alias{write_explanation}
\title{Writes explanation to the .Rmd report files.}
\usage{
write_explanation()
}
\value{
string of the explanation for rmd file
}
\description{
Writes explanation to the .Rmd report files.
}

