% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_head}
\alias{write_head}
\title{Writes head of the .Rmd report files.}
\usage{
write_head(title)
}
\arguments{
\item{title}{(string) title to include in header}
}
\value{
string of the header for rmd file
}
\description{
Writes head of the .Rmd report files.
}
