% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_unassigned_reads}
\alias{write_unassigned_reads}
\title{Writes unassigned reads}
\usage{
write_unassigned_reads(barcode, top = 5)
}
\arguments{
\item{barcode}{(string) CSV name of the barcode unassigned reads file.}

\item{top}{(numeric) How many from the most abundant unassigned reads write.}
}
\value{
string to include in rmd file
}
\description{
Writes unassigned reads
}

