% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_helpers.R
\name{writeConfigError}
\alias{writeConfigError}
\title{This function write info into a file that is human readable. It also tells
the users via console that there was an error, and where he can find further
verbose information.}
\usage{
writeConfigError(currentConfigPath, logFileName, logFileConn)
}
\arguments{
\item{(string)}{currentConfigPath: The place where is the original config file.}

\item{(string)}{logFileName:       The path where the log file is placed.
Usually, the name goes like this:
    resultsFolder + "/MasterLog.txt"}

\item{(file)}{logFileConn:       The file object that refers to "logFileName".}
}
\value{
(void)
}
\description{
This function write info into a file that is human readable. It also tells
the users via console that there was an error, and where he can find further
verbose information.
}

