% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{xlabels_spacing}
\alias{xlabels_spacing}
\title{Creates equal label spacing.
Used to calculate x label ticks.}
\usage{
xlabels_spacing(direction, box, ampl_len, spacing)
}
\arguments{
\item{direction}{(numeric) 0 or 1}

\item{box}{(IRanges) specifies where predcited cut site is}

\item{ampl_len}{(numeric) Length of the amplicon}

\item{spacing}{(numeric) desired spacing between ticks}
}
\value{
numeric vector
}
\description{
Creates equal label spacing.
Used to calculate x label ticks.
}
