% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_filters.R
\name{alphabetQuality}
\alias{alphabetQuality}
\title{This filters out sequences which have nonstandard nucleotides.}
\usage{
alphabetQuality(reads)
}
\arguments{
\item{reads}{(ShortRead object) Loaded reads from fastq.}
}
\value{
(boolean) Logical vector with the valid rows as TRUE.
}
\description{
This filters out sequences which have nonstandard nucleotides.
}
\keyword{internal}
