% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplican.R
\docType{package}
\name{amplican}
\alias{amplican-package}
\alias{amplican}
\title{Automated analysis of CRISPR experiments.}
\description{
Main goals:
\enumerate{
\item Flexible pipeline for analysis of the CRISPR Mi-Seq or Hi-Seq data.
\item Compatible with GRanges and data.table style.
\item Precise quantification of mutation rates.
\item Prepare automatic reports as .Rmd files that are flexible
and open for manipulation.
\item Provide specialized plots for deletions, insertions, mismatches,
variants, heterogeneity of the reads.
}
}
\details{
To learn more about amplican, start with the vignettes:
\code{browseVignettes(package = "amplican")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/valenlab/amplican}
  \item Report bugs at \url{https://github.com/valenlab/amplican/issues}
}

}
\author{
\strong{Maintainer}: Eivind Valen \email{eivind.valen@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Kornel Labun \email{kornel.labun@gmail.com}
}

}
