% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{amplican_print_reads}
\alias{amplican_print_reads}
\title{Pretty print forward and reverse reads aligned to each other.}
\usage{
amplican_print_reads(forward, reverse)
}
\arguments{
\item{forward}{(character or vector of characters) Forward reads.}

\item{reverse}{(character or vector of characters) Will be reverse
complemented before alignment.}
}
\value{
Vector with alignments ready to be printed.
}
\description{
Usefull and needed for barcode reports.
}

