% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanNormalize.R
\name{amplicanNormalize}
\alias{amplicanNormalize}
\title{Remove events that can be found in Controls.}
\usage{
amplicanNormalize(aln, cfgT, add = c("guideRNA", "Group"),
  skip = c("counts", "score"))
}
\arguments{
\item{aln}{(data.frame) Contains events from alignments.}

\item{cfgT}{(data.frame) Config table with information about experiments.}

\item{add}{(character vector) Columns from cfgT that should be included
in event table for normaliztion matching. Defaults to c("guideRNA", "Group"),
which means that only those events created by the same guideRNA in the same
Group will be removed if found in Control.}

\item{skip}{(character vector) Specifies which column of aln to skip,
defaults to c('counts', 'score').}
}
\value{
(data.frame) Same as aln, but events are normalized. Events from
Control are not changed. Additionaly columns from add are added to the
data.frame.
}
\description{
This function can adjust events for small differences between
known annotations (amplicon sequences) and real DNA of the strain that
was sequenced.
}
\examples{
aln <- data.frame(seqnames = 1:5,
                  start = 1,
                  end = 2,
                  width = 2,
                  counts = 101:105)
cfgT <- data.frame(ID = 1:5,
                   guideRNA = rep("ACTG", 5),
                   Group = c("A", "A", "B", "B", "B"),
                   Control = c(TRUE, FALSE, TRUE, FALSE, FALSE))
# all events are same as in the control group, therfore are filtered out
# events from control groups stay
amplicanNormalize(aln, cfgT)
# events that are different from control group are preserved
aln[2, "start"] <- 3
amplicanNormalize(aln, cfgT)

}
\seealso{
Other analysis steps: \code{\link{amplicanAlign}},
  \code{\link{amplicanFilter}},
  \code{\link{amplicanOverlap}},
  \code{\link{amplicanPipeline}},
  \code{\link{amplicanReport}},
  \code{\link{amplicanSummarize}},
  \code{\link{map_to_relative}}
}
