% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplican.R
\name{amplicanPipeline}
\alias{amplicanPipeline}
\title{Wraps main package functionality into one function.}
\usage{
amplicanPipeline(config, fastq_folder, results_folder, total_processors = 1,
  min_quality = 0, fastqfiles = 2, PRIMER_DIMER = 30, cut_buffer = 5)
}
\arguments{
\item{config}{(string) The path to your configuration file. For example:
\code{system.file('extdata', 'config.txt', package = 'amplican')}}

\item{fastq_folder}{(string) Path to FASTQ files. If not specified,
FASTQ files should be in the same directory as config file.}

\item{results_folder}{(string) Where do you want your results to be stored.
The package will create files in that folder so make sure you have writing
permissions.}

\item{total_processors}{(numeric) Set this to the number of processors you
want to use. Default is 1. Works only if you have 'doParallel' installed
and accessible.}

\item{min_quality}{(numeric)  Similar as in average_quality, but this is
the minimum quality for ALL nucleotides. If one of them has quality BELLOW
this threshold, then the sequence is skipped. Default is 0.}

\item{fastqfiles}{(numeric) Normally you want to use both FASTQ files. But in
some special cases, you may want to use only the
forward file, or only the reverse file.
0 - Use both FASTQ files
0.5 - Use both FASTQ files,
but only for one of the reads (forward or reverse)
is required to have primer perfectly matched to
sequence - eg. use when reverse reads are
trimmed of primers, but forward reads have
forward primer in the sequence
1 - Use only the forward FASTQ file
2 - Use only the reverse FASTQ file}

\item{PRIMER_DIMER}{(numeric) Value specyfying buffer for PRIMER DIMER
detection. For a given read it will be recognized as PRIMER DIMER when
alignment will introduce gap of size bigger than:
length of amplicon - (lenghts of PRIMERS + PRIMER_DIMER value)}

\item{cut_buffer}{(numeric) Value specyfying a buffer for PAM, this will add
from both sides to a window defined from uppercase letters in the amplicon.
Deletions overlapping this window will be considered a
valid cut (if confirmed by both forward and rewerse reads).}
}
\value{
NULL All results are created in specified results_folder.
}
\description{
amplican_pipeline is convinient wrapper around all most popular settings of
\code{\link{amplicanAnalysis}} and \code{\link{amplicanReport}}.
It will generate all results in the \code{result_folder} and also print
prepared reports into 'reports' folder.
}
\examples{
config <- system.file("extdata", "config.csv", package = "amplican") #example config file
fastq_folder <- system.file("extdata", "", package = "amplican") #path to example fastq files
results_folder <- paste0(fastq_folder, "results") #output folder
amplicanPipeline(config, fastq_folder, results_folder)

}
\seealso{
Other analysis steps: \code{\link{amplicanAnalysis}},
  \code{\link{amplicanReport}}
}

