% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanReport.R
\name{amplicanReport}
\alias{amplicanReport}
\title{Prepare report as .Rmd file}
\usage{
amplicanReport(results_folder, levels = c("id", "barcode", "experiment_type",
  "guide"), report_files = c("report_id", "report_barcode",
  "report_experiment", "report_guide"))
}
\arguments{
\item{results_folder}{(string) Folder containing results from the \code{\link{amplicanAnalysis}} function,
do not change names of the files.}

\item{levels}{(vector) Possible values are: "id", "barcode", "experiment_type", "guide". You can also input more than one
value eg. c("id", "barcode") will create two separate reports for each level.}

\item{report_files}{(vector) You can supply your own names of the files. For each of the levels there has to be one
file name. Files are created in current working directory by default.}
}
\description{
amplicanReport takes a configuration file, fastq reads and output directory to prepare
summary as editable .Rmd file. You can specify whether you want to make summaries based on
ID, Barcode, Experiment Type or even guideRNA.
}

