% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplicanReport.R
\name{ampliCanReport}
\alias{ampliCanReport}
\title{Prepare report as .Rmd file}
\usage{
ampliCanReport(alignments_folder, levels = c("ID"),
  report_files = c("report_id"))
}
\arguments{
\item{alignments_folder}{(string) Folder containing results from the \code{\link{ampliCanAnalysis}} function,
do not change names of the files.}

\item{levels}{(vector) Possible values are: "ID", "Barcode", "Experiment_Type". You can also input more than one
value eg. c("ID", "Barcode") will create two separate reports for each level.}

\item{report_files}{(vector) You can supply your own names of the files. For each of the levels there has to be one
file name. Files are created in current working directory by default.}
}
\description{
ampliCanReport takes a configuration file, fastq reads and output directory to prepare
summary as editable .Rmd file. You can specify whether you want to make summaries based on
ID, Barcode or Experiment Type.
}

