% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{assignedCount}
\alias{assignedCount}
\title{Get count of assigned reads.}
\usage{
assignedCount(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Get count of assigned reads.
}
