% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{barcodeData<-}
\alias{barcodeData<-}
\title{Barcode data.}
\usage{
barcodeData(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(data.frame)}
}
\description{
Set barcode data.frame with information on the barcode level.
}
