% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{barcodeData}
\alias{barcodeData}
\title{Barcode data.}
\usage{
barcodeData(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(data.tableOrNULL)
}
\description{
Get barcode data.frame with information on the barcode level.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "AlignmentsExperimentSet.rds", package = "amplican")
aln <- readRDS(file_path)
barcodeData(aln)
}
\keyword{internal}
