% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{checkConfigFile}
\alias{checkConfigFile}
\title{This function pre-process a config file and checks that everything is in
order.}
\usage{
checkConfigFile(configTable, fastq_folder)
}
\arguments{
\item{configTable}{(data.frame) Config file.}

\item{fastq_folder}{(string) Path to fastq folder.}
}
\value{
(boolean) TRUE, If anything goes wrong stops and prints error.
}
\description{
Its takes care of the following:
  No IDs are duplicated.
  Every combination of barcode, forward primer and reverse primer is unique.
  Each barcode has unique forward reads file and reverse read files.
  Checks that the read files exist with read access.
}

