% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{checkFileAccess}
\alias{checkFileAccess}
\title{This function check if the given file exist and can be read.}
\usage{
checkFileAccess(filePath)
}
\arguments{
\item{filePath}{(String) A string the path to the file.}
}
\value{
(Void) Stop if no access.
}
\description{
This function check if the given file exist and can be read.
}

