% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_directory.R
\name{checkFileWriteAccess}
\alias{checkFileWriteAccess}
\title{This function checks if the given directory exist and can be written to.}
\usage{
checkFileWriteAccess(filePath)
}
\arguments{
\item{filePath}{(string) A string the path to the file.}
}
\value{
(invisible) TRUE, Stop if no access.
}
\description{
This function checks if the given directory exist and can be written to.
}

