% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{checkPositions}
\alias{checkPositions}
\title{This function checks if the given alignment positions are valid}
\usage{
checkPositions(alignmentPositions, amplicon, ID, barcode, logFileConn)
}
\arguments{
\item{alignmentPositions}{(vector) of integers representing the aligning positions inside the amplicon}

\item{amplicon}{(string) A sequence of nucleotides in a string format representing the amplicon}

\item{ID}{(string) The ID for target and amplicon.}

\item{barcode}{(string) barcode belonging to target and amplicon.}

\item{logFileConn}{(connection) Path to config file.}
}
\value{
(boolean) TRUE when positions are greater than 0.
}
\description{
This function checks if the given alignment positions are valid
}

