% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{checkPrimers}
\alias{checkPrimers}
\title{This function checks if the forward and reverse primer are in the amplicon
and where they are located.}
\usage{
checkPrimers(configTable, fastqfiles)
}
\arguments{
\item{configTable}{(data.frame) A data frame of config file.}

\item{fastqfiles}{(numeric) Which primers are important.}
}
\value{
configTable (data.frame) A data frame of config file with additional
fields for start locations of the primers
}
\description{
This function checks if the forward and reverse primer are in the amplicon
and where they are located.
}

