% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_helpers.R
\name{checkTarget}
\alias{checkTarget}
\title{This function check if the target RNA is in the amplicon.}
\usage{
checkTarget(targetPrimer, amplicon, ID, barcode, logFileConn)
}
\arguments{
\item{targetPrimer}{(string) A sequence of nucleotides in a string format representing the target}

\item{amplicon}{(string) A sequence of nucleotides in a string format representing the amplicon}

\item{ID}{(string) The ID from where this target and amplicon came.}

\item{barcode}{(String) The barcode from where this target and amplicon came.}

\item{logFileConn}{(connection) A file descriptor in R which is pointing to the log file.}
}
\value{
(bool) , TRUE:  Everything went good.
}
\description{
This function check if the target RNA is in the amplicon.
}

