% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_warnings.R
\name{checkTarget}
\alias{checkTarget}
\title{Checks if the guideRNA is in the amplicon.}
\usage{
checkTarget(configTable)
}
\arguments{
\item{configTable}{(data.frame) data frame of config file}
}
\value{
(boolean vector) Prints warning when some guides can't be found.
}
\description{
Checks if the guideRNA is in the amplicon.
}
