% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_directory.R
\name{deleteFiles}
\alias{deleteFiles}
\title{Remove forward and reverse fastq files.}
\usage{
deleteFiles(configTable)
}
\arguments{
\item{configTable}{(data.frame) Contains Forward_Reads_File and
Reverse_Reads_File to be removed}
}
\value{
(void) In case of fail, prints err.
}
\description{
Delete files from configTables Forward_Reads_File and
Reverse_Reads_File (fastq files).
}

