% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{experimentData<-}
\alias{experimentData<-}
\title{Experiment data.}
\usage{
experimentData(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(data.frame)}
}
\description{
Set experiment data.frame with information on the experiment level.
}
