% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{experimentData}
\alias{experimentData}
\title{Experiment data.}
\usage{
experimentData(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(data.frameOrNULL)
}
\description{
Get experiment data.frame with information on the experiment level.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "AlignmentsExperimentSet.rds", package = "amplican")
aln <- readRDS(file_path)
experimentData(aln)
}
\keyword{internal}
