% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{extractEvents}
\alias{extractEvents}
\title{Extract AlignmentsExperimentSet events into data.frame.}
\usage{
extractEvents(object, total_processors = 1)
}
\arguments{
\item{object}{(AlignmentsExperimentSet)}

\item{total_processors}{(numeric) Number of cores to use, defaults to 1.}
}
\value{
(data.frame) Compatible with \code{\link[GenomicRanges]{GRanges}}
style.
}
\description{
Extracts events (insertions, deletions, mismatches) from alignments into
data.frame. Can use multiple cores as process is quite slow. All events are
relative towards forward strand. "-" in strand column indicates which events
were from reverse reads.
}
