% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_alignment.R
\name{flipRanges}
\alias{flipRanges}
\title{Reverse complement events that have amplicons with direction 1.}
\usage{
flipRanges(idR, cfgT)
}
\arguments{
\item{idR}{(data.frame) Loaded events.}

\item{cfgT}{(data.frame) Loaded configuration file.}
}
\value{
(data.frame) Returns input idR, but events for amplicons with
direction 1 reverse complemented, "+" and "-" swapped.
}
\description{
Reverse complement events that have amplicons with direction 1.
}
