% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{fwdReads}
\alias{fwdReads}
\title{Alignments for forward reads.}
\usage{
fwdReads(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Get alignments for forward reads.
}
