% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{get_left_primer}
\alias{get_left_primer}
\title{left primer sequence}
\usage{
get_left_primer(config, id)
}
\arguments{
\item{config}{(data.frame) config table}

\item{id}{(vector) a vector of id's}
}
\value{
(character) left primer sequence
}
\description{
left primer sequence
}
