% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plots.R
\name{get_right_primer}
\alias{get_right_primer}
\title{right primer sequence}
\usage{
get_right_primer(config, id)
}
\arguments{
\item{config}{(data.frame) config table}

\item{id}{(vector) a vector of id's}
}
\value{
(character) right primer sequence
}
\description{
right primer sequence
}
