% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_helpers.R
\name{getEventInfo}
\alias{getEventInfo}
\title{This function takes a lite string and gives back the events coordinates}
\usage{
getEventInfo(liteString)
}
\arguments{
\item{liteString}{(list)}
}
\value{
A list with 6 elements:
    - The element 1 and element 4 is the total of insertions and deletions (in that order)
    - The element 2 and element 5 is where the insertions and deletions starts (in that order)
    - The element 3 and element 6 is where the insertions and deletion ends in (that order)
    - Elements 2,3,4,5 are integer arrays.
    - If the total of insertions/deletions is 0, the integer arrays will be equal to NULL
}
\description{
For the given example:
0         1         2         3         4         5
012345678901234567890123456789012345678901234567890123456789
NNNNNNNNNNN-----NNNNNNN-------NNNNNNNN--------NNNNNNNN------    Subject
------NNNNNNNNNNNNNNNNNNNNNNNNN---NNNNNNNNNNNNN---NNNNNNNNNN    Pattern
}
\details{
The RCPP library returns the information regarding the start and ends of the events in the
following way:
(Please notice that there is a +1 shift in the index)

#returned: @4@ 12,16 * 24,30 * 39,46 * 55,60 * ! @3@ 1,6 * 31,33  * 48,50 *
}

