% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_helpers.R
\name{makeAlignment}
\alias{makeAlignment}
\title{make alignments}
\usage{
makeAlignment(SKIP_BAD_NUCLEOTIDES = TRUE, AVERAGE_QUALITY = 0,
  MIN_QUALITY = 0, WRITE_ALIGNMENTS = TRUE, SCORING_MATRIX = "NUC44",
  GAP_OPENING = 50, GAP_EXTENSION = 0, GAP_ENDING = FALSE,
  FAR_INDELS = TRUE, configDataframe = NULL, resultFolder = NULL,
  alignmentFolder = NULL, processID = 0, configFilePath = "", TIMING,
  TEMPFOLDER, FASTQFILES)
}
\arguments{
\item{AVERAGE_QUALITY}{(int)}

\item{MIN_QUALITY}{(int)}

\item{WRITE_ALIGNMENTS}{(bool)}

\item{SCORING_MATRIX}{(char)}

\item{GAP_OPENING}{(bool)}

\item{GAP_EXTENSION}{(int)}

\item{GAP_ENDING}{(bool)}

\item{FAR_INDELS}{(bool)}

\item{configDataframe}{(bool)}

\item{resultFolder}{(char)}

\item{alignmentFolder}{(char)}

\item{configFilePath}{(char)}

\item{TIMING}{(bool)}

\item{TEMPFOLDER}{(char)}

\item{FASTQFILES}{(char)}

\item{SKIP_BAD_NUKLEOTYDES}{(bool)}

\item{procesID}{(int)}
}
\value{
no clue
}
\description{
make alignments
}

