% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_alignment.R
\name{makeAlignment}
\alias{makeAlignment}
\title{Make alignments}
\usage{
makeAlignment(configTable, resultsFolder, skip_bad_nucleotides = TRUE,
  average_quality = 0, min_quality = 0, write_alignments = 1,
  scoring_matrix = "NUC44", gap_opening = 50, gap_extension = 0,
  gap_ending = FALSE, far_indels = TRUE, fastqfiles = 0,
  PRIMER_DIMER = 10, cut_buffer = 5)
}
\arguments{
\item{configTable}{(bool)}

\item{resultsFolder}{(char)}

\item{skip_bad_nucleotides}{(bool)}

\item{average_quality}{(int)}

\item{min_quality}{(int)}

\item{write_alignments}{(bool)}

\item{scoring_matrix}{(char)}

\item{gap_opening}{(bool)}

\item{gap_extension}{(int)}

\item{gap_ending}{(bool)}

\item{far_indels}{(bool)}

\item{fastqfiles}{(char)}

\item{PRIMER_DIMER}{(numeric)}

\item{cut_buffer}{(numeric)}
}
\value{
Void
}
\description{
Make alignments
}

