% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_alignment.R
\name{makeAlignment}
\alias{makeAlignment}
\title{Make alignments helper.}
\usage{
makeAlignment(configTable, resultsFolder, skip_bad_nucleotides, average_quality,
  min_quality, write_alignments, scoring_matrix, gap_opening, gap_extension,
  fastqfiles, PRIMER_DIMER, cut_buffer)
}
\arguments{
\item{configTable}{config file as data table}

\item{resultsFolder}{(string) path to resultsFolder}

\item{skip_bad_nucleotides}{(logical) Some sequences have faulty nucleotides
labels with N. If we find a sequence like that in either forwards or
reverse, we skip that alignment. Default is TRUE.}

\item{average_quality}{(numeric) The FASTQ file have a quality for each
nucleotide, being ! the lower and ~ the highest. In ASCII :
                             !'#$%&'()*+,-./0123456789:;<=>?@
                             ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_`
                             abcdefghijklmnopqrstuvwxyz{|}~

This quality variable goes from 0 to 100, being 0 the lowest quality and
100 the highest. You can write whatever number in between, and the program
will find out the appropriate character encoding. The filter works by
converting each character to a number, and then finding the average. If the
average fall above this threshold then we take the sequence. Default is 0.}

\item{min_quality}{(numeric)  Similar as in average_quality, but this is
the minimum quality for ALL nucleotides. If one of them has quality BELLOW
this threshold, then the sequence is skipped. Default is 0.}

\item{write_alignments}{(boolean) Whether we should write alignments results
to separate files for each ID}

\item{scoring_matrix}{(string) For now the only option is 'NUC44'.}

\item{gap_opening}{(numeric) The opening gap score. Default is 50.}

\item{gap_extension}{(numeric) The gap extension score. Default is 0.}

\item{fastqfiles}{(numeric) Normally you want to use both FASTQ files. But in
some special cases, you may want to use only the
forward file, or only the reverse file.
0 - Use both FASTQ files
0.5 - Use both FASTQ files,
but only for one of the reads (forward or reverse)
is required to have primer perfectly matched to
sequence - eg. use when reverse reads are
trimmed of primers, but forward reads have
forward primer in the sequence
1 - Use only the forward FASTQ file
2 - Use only the reverse FASTQ file}

\item{PRIMER_DIMER}{(numeric) Value specifying buffer for PRIMER DIMER
detection. For a given read it will be recognized as PRIMER DIMER when
alignment will introduce gap of size bigger than:
length of amplicon - (lengths of PRIMERS + PRIMER_DIMER value)}

\item{cut_buffer}{(numeric) Value specifying a buffer for PAM, this will add
from both sides to a window defined from uppercase letters in the amplicon.
Deletions overlapping this window will be considered a
valid cut (if confirmed by both forward and reverse reads).}
}
\value{
resultsFolder as invisible
}
\description{
Main functionality of the package, aligning reads to the amplicon.
}
