% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{map_to_relative}
\alias{map_to_relative}
\title{Map events to their respective relative coordinates specified with
UPPER case.}
\usage{
map_to_relative(events, config)
}
\arguments{
\item{events}{(data.frame) List of events to map to the relative coordinates.}

\item{config}{(data.frame) config table}
}
\value{
(GRanges) Same as events, but the coordinates are relative to the
}
\description{
Translate coordinates of GRanges events so that they can be relative to the
amplicon. As point zero we assume first left sided UPPER case letter in the
amplicon.
}
\examples{
# example config
config <- read.csv(system.file("extdata", "config.csv",
                   package = "amplican"))
# example events
events <- read.csv(system.file("extdata", "results",
                   "alignments_events.csv", package = "amplican"))
# make events relative to the UPPER case
map_to_relative(events, config)

}
