% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{plot_height}
\alias{plot_height}
\title{Function to calculate figure height
based on number of elements to plot.}
\usage{
plot_height(x)
}
\arguments{
\item{x}{(numeric) number of elements to fit onto height axis}
}
\value{
(numeric)
}
\description{
Function to calculate figure height
based on number of elements to plot.
}

