% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{readCounts}
\alias{readCounts}
\title{Alignments for forward reads.}
\usage{
readCounts(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Set alignments for forward reads.
}
