% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{rveReads<-}
\alias{rveReads<-}
\title{Alignments for forward reads.}
\usage{
rveReads(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(list) Named (experiment IDs) list with elements of}
}
\value{
(AlignmentsExperimentSet)
\code{\link{PairwiseAlignmentsSingleSubject}} class.
}
\description{
Set alignments for forward reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "AlignmentsExperimentSet.rds", package = "amplican")
aln <- readRDS(file_path)
rveReads(aln) <- rveReads(aln) # replace with the same values
}
\keyword{internal}
