% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{rveReads}
\alias{rveReads}
\title{Alignments for reverse reads.}
\usage{
rveReads(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Get alignments for reverse reads.
}
