% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{rveReadsType}
\alias{rveReadsType}
\title{Type of reverse reads.}
\usage{
rveReadsType(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(listOrNULL) list with objects of PairwiseAlignmentsSingleSubject
}
\description{
Get type of reverse reads.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "AlignmentsExperimentSet.rds", package = "amplican")
aln <- readRDS(file_path)
rveReadsType(aln)
}
\keyword{internal}
