% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_filters.R
\name{thresholdScores}
\alias{thresholdScores}
\title{Find Off-targets and Fragmented alignments from reads.}
\usage{
thresholdScores(x)
}
\arguments{
\item{x}{(numeric) Input alignment scores of all reads. Should work on
scores from single experiment.}
}
\value{
(numeric) Threshold value for suggested cutoff on the scores of the
alignments. Values below this threshold are potential off-targets or low
quality alignments.
}
\description{
Will try to detect off-targets and low quality alignments (outliers). It
checks for continuity of scores and has no asumptions about underlying
distributions. When there is gap in scores that is bigger
than expected, will return threshold. When there is less than 1000
scores in \code{x} minimum score is returned.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "raw_events.csv", package = "amplican")
aln <- data.table::fread(file_path)
thresholdScores(aln$score[aln$seqnames == "ID_1"])

}
\seealso{
\code{\link{findPD}} \code{\link{findEOP}}

Other filters: \code{\link{findEOP}}, \code{\link{findPD}}
}
