% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{unassignedCount}
\alias{unassignedCount}
\title{Get count of unassigned reads.}
\usage{
unassignedCount(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Get count of unassigned reads.
}
