% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{unassignedData<-}
\alias{unassignedData<-}
\title{Alignments for forward reads.}
\usage{
unassignedData(x) <- value
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{value}{(list) Named (experiment IDs) list with elements of
\code{\link[Biostrings]{PairwiseAlignmentsSingleSubject}} class.}
}
\description{
Set alignments for forward reads.
}
