% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{unassignedData}
\alias{unassignedData}
\title{Unassigned read information.}
\usage{
unassignedData(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\description{
Get unassigned reads and their characteristics.
}
