% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{unassignedData}
\alias{unassignedData}
\title{Unassigned read information.}
\usage{
unassignedData(x)
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}
}
\value{
(data.frameOrNULL)
}
\description{
Get unassigned reads and their characteristics.
}
\examples{
file_path <- system.file("extdata", "results", "alignments",
                         "AlignmentsExperimentSet.rds", package = "amplican")
aln <- readRDS(file_path)
unassignedData(aln)
}
\keyword{internal}
