% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_directory.R
\name{unifyFiles}
\alias{unifyFiles}
\title{This function takes all the files in a given folder that match a regex
pattern and merges them together.}
\usage{
unifyFiles(targetFolder, regex, finalFileName, header = T, delete = T,
  isrecursive = F)
}
\arguments{
\item{targetFolder}{(string) The folder where the files are suppose to be.}

\item{regex}{(string) The files to be combined, must comply with this regex.}

\item{finalFileName}{(String) The name of the final file.}

\item{header}{(boolean) If header is true, it will take the header of the
first file only and ignore the rest. Otherwise it
will repeat the header many times.}

\item{delete}{(boolean) If true, it will delete the original files}

\item{isrecursive}{(boolean) If TRUE will search recursively through regex matching folders.}
}
\value{
(void)
}
\description{
This function takes all the files in a given folder that match a regex
pattern and merges them together.
}

