% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_helpers.R
\name{unifyFiles}
\alias{unifyFiles}
\title{This function takes all the files in a given folder that match a regex
pattern and put them toguether into a single file.}
\usage{
unifyFiles(targetFolder, regex, finalFileName, header, delete, isrecursive)
}
\arguments{
\item{targetFolder:}{(String) The folder where the files are suppose to be.}

\item{regex:}{(String) The files to be combined, must comply with this
regex.}

\item{finalFileName:}{(String) The name of the final file.}

\item{header:}{(Bool)   If header is true, it will take the header of the
first file only and ignore the rest. Otherwise it
will repeat the header many times.}

\item{delete:}{(Bool)   If true, it will delete the original files}

\item{isrecursive:}{(Bool)   If recursive is true, it will look inside all the
folder of the given folder for more files which
name match the regex.}
}
\value{
(String) A string with the nucleotides reversed and complemented
}
\description{
BE CAREFULL: This function __WILL_DELETE__ all the original little files
afterward if one of the parameters it set to TRUE.
}
\details{
The function takes the following parameters:
}

