% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_helpers.R
\name{upperGroups}
\alias{upperGroups}
\title{For a given string, detect how many groups of uppercases are there, where are
they, and how long they are.}
\usage{
upperGroups(candidate)
}
\arguments{
\item{candidate}{(String) A string with the nucleotide sequence.}
}
\value{
(Ranges) A Ranges object with uppercases groups for given candidate string
}
\description{
For example:
   asdkfaAGASDGAsjaeuradAFDSfasfjaeiorAuaoeurasjfasdhfashTTSfajeiasjsf
}
\details{
Has 4 groups of uppercases of length 7, 4, 1 and 3.
}

