% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rmd.R
\name{write_alignment_plots}
\alias{write_alignment_plots}
\title{Writes alignment plots - aggregate on the amplicon: deletion, insertion and mismatch.}
\usage{
write_alignment_plots(ID, title, cut_buffer = 5)
}
\arguments{
\item{ID}{(string or vector of strings)}

\item{title}{(string) title for the alignment plots}

\item{cut_buffer}{(numeric) cut detection box will be extended left and right by box_buffer}
}
\value{
string to include in rmd file
}
\description{
Writes alignment plots - aggregate on the amplicon: deletion, insertion and mismatch.
}

