% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignmentsExperimentSet-class.R
\name{writeAlignments}
\alias{writeAlignments}
\title{Alignments for forward reads.}
\usage{
writeAlignments(x, file = "", aln_format = "txt")
}
\arguments{
\item{x}{(AlignmentsExperimentSet)}

\item{file}{(connection or string) Destination file. When empty, defaults to
standard output.}

\item{aln_format}{("txt" or "fasta") Specifies format of the file.}
}
\description{
Set alignments for forward reads.
}
